Rename SmallRye Fault Tolerance 'executor' module to 'deployment' and fix package name accordingly.
