+++ boms/standard-expansion/pom.xml
@@ -95,7 +95,7 @@
 
             <dependency>
                 <groupId>${full.maven.groupId}</groupId>
-                <artifactId>wildfly-microprofile-fault-tolerance-smallrye-executor</artifactId>
+                <artifactId>wildfly-microprofile-fault-tolerance-smallrye-deployment</artifactId>
                 <version>${project.version}</version>
                 <exclusions>
                     <exclusion>
