+++ microprofile/fault-tolerance-smallrye/deployment/src/main/java/org/wildfly/microprofile/faulttolerance/deployment/FaultToleranceContainerExecutorFactory.java
@@ -20,7 +20,7 @@
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
 
-package org.wildfly.extension.microprofile.faulttolerance;
+package org.wildfly.microprofile.faulttolerance.deployment;
 
 import java.util.OptionalInt;
 import java.util.concurrent.ThreadFactory;
