+++ microprofile/galleon-common/src/main/resources/license/microprofile-feature-pack-licenses.xml
@@ -679,7 +679,7 @@
     </dependency>
     <dependency>
       <groupId>${full.maven.groupId}</groupId>
-      <artifactId>wildfly-microprofile-fault-tolerance-smallrye-executor</artifactId>
+      <artifactId>wildfly-microprofile-fault-tolerance-smallrye-deployment</artifactId>
       <licenses>
         <license>
           <name>GNU Lesser General Public License v2.1 or later</name>
