+++ microprofile/galleon-common/src/main/resources/modules/system/layers/base/io/smallrye/fault-tolerance/main/module.xml
@@ -49,6 +49,6 @@
         <module name="org.jboss.logging"/>
         <module name="org.jboss.weld.core"/>
         <module name="org.jboss.weld.spi"/>
-        <module name="org.wildfly.microprofile.fault-tolerance-smallrye.executor"/>
+        <module name="org.wildfly.microprofile.fault-tolerance-smallrye.deployment"/>
     </dependencies>
 </module>
