+++ microprofile/galleon-common/src/main/resources/modules/system/layers/base/org/wildfly/microprofile/fault-tolerance-smallrye/deployment/main/module.xml
@@ -22,14 +22,14 @@
   ~ 02110-1301 USA, or see the FSF site: http://www.fsf.org.
   -->
 
-<module xmlns="urn:jboss:module:1.9" name="org.wildfly.microprofile.fault-tolerance-smallrye.executor">
+<module xmlns="urn:jboss:module:1.9" name="org.wildfly.microprofile.fault-tolerance-smallrye.deployment">
 
     <properties>
         <property name="jboss.api" value="private"/>
     </properties>
 
     <resources>
-        <artifact name="${org.wildfly:wildfly-microprofile-fault-tolerance-smallrye-executor}"/>
+        <artifact name="${org.wildfly:wildfly-microprofile-fault-tolerance-smallrye-deployment}"/>
     </resources>
 
     <dependencies>
