+++ testsuite/layers/src/test/java/org/jboss/as/test/layers/LayersTestCase.java
@@ -230,7 +230,7 @@ public class LayersTestCase {
                     "io.smallrye.fault-tolerance",
                     "org.eclipse.microprofile.fault-tolerance.api",
                     "org.wildfly.extension.microprofile.fault-tolerance-smallrye",
-                    "org.wildfly.microprofile.fault-tolerance-smallrye.executor",
+                    "org.wildfly.microprofile.fault-tolerance-smallrye.deployment",
                     // Used by Hibernate Search but only in preview TODO this doesn't seem right; NOT_REFERENCED should suffice
                     "org.hibernate.search.mapper.orm.coordination.outboxpolling"
             );
